switched the ordering of the tools so we can turn on and off the rotator
