+++ SmartSlicePlugin/select_tool/LoadHandle.py
@@ -1,6 +1,7 @@
 import math
 
 from typing import Optional
+from enum import Enum
 
 from UM.Math.Color import Color
 from UM.Math.Vector import Vector
@@ -9,72 +10,126 @@ from UM.Mesh.MeshBuilder import MeshBuilder
 from UM.Scene.ToolHandle import ToolHandle
 from UM.Scene.SceneNode import SceneNode
 
-from .LoadArrow import LoadArrow
 from ..utils import angleBetweenVectors
+from .LoadRotator import LoadRotator
 
 
 class LoadHandle(ToolHandle):
-    """Provides the circular toolhandle and arrow for the load direction"""
+    """Provides the arrow for the load direction"""
 
-    color = Color(0.4, 0.4, 1., 1.)
+    color = LoadHandle.color
 
-    INNER_RADIUS = 2.0 * LoadArrow.ARROW_TOTAL_LENGTH
-    LINE_WIDTH = 1.0
-    OUTER_RADIUS = INNER_RADIUS + LINE_WIDTH
-    ACTIVE_INNER_RADIUS = INNER_RADIUS - 3
-    ACTIVE_OUTER_RADIUS = OUTER_RADIUS + 3
-    ACTIVE_LINE_WIDTH = LINE_WIDTH + 3
+    ARROW_HEAD_LENGTH = 8
+    ARROW_TAIL_LENGTH = 22
+    ARROW_TOTAL_LENGTH = ARROW_HEAD_LENGTH + ARROW_TAIL_LENGTH
+    ARROW_HEAD_WIDTH = 2.8
+    ARROW_TAIL_WIDTH = 0.8
 
-    def __init__(self, parent = None):
+    def __init__(self, parent = None, start: Vector = Vector(0, 0, 0), direction: Vector = Vector.Unit_X):
         super().__init__(parent)
 
-        self._name = "LoadHandle"
+        self._name = "LoadArrow"
         self._auto_scale = False
 
-        self._center = Vector(0, 0, 0)
-        self._rotation_axis = Vector.Unit_Z
-        self._arrow_direction = Vector.Unit_X
-
-        self.show_rotation = True
-
-        self._arrow = LoadArrow(self, self._center, self._arrow_direction)
+        self.start = start
+        self.direction = direction
+        self._handle = LoadRotator(self)
 
     def buildMesh(self):
+
         mb = MeshBuilder()
 
-        if self.show_rotation:
+        self.direction = self.direction.normalized()
 
-            #SOLIDMESH
-            mb.addDonut(
-                inner_radius = self.INNER_RADIUS,
-                outer_radius = self.OUTER_RADIUS,
-                width = self.LINE_WIDTH,
-                axis = self._rotation_axis,
-                color = self.color
-            )
+        p_base0 = Vector(
+            self.start.x + self.direction.x * self.ARROW_HEAD_LENGTH,
+            self.start.y + self.direction.y * self.ARROW_HEAD_LENGTH,
+            self.start.z + self.direction.z * self.ARROW_HEAD_LENGTH
+        )
+        p_tail0 = Vector(
+            self.start.x + self.direction.x * self.ARROW_TOTAL_LENGTH,
+            self.start.y + self.direction.y * self.ARROW_TOTAL_LENGTH,
+            self.start.z + self.direction.z * self.ARROW_TOTAL_LENGTH
+        )
 
-            self.setSolidMesh(mb.build())
+        p_base0 = Vector(
+            self.start.x + self.direction.x * self.ARROW_TAIL_LENGTH,
+            self.start.y + self.direction.y * self.ARROW_TAIL_LENGTH,
+            self.start.z + self.direction.z * self.ARROW_TAIL_LENGTH
+        )
+        p_head = p_tail0
+        p_tail0 = self.start
 
-            #SELECTIONMESH
-            mb.addDonut(
-                inner_radius = self.ACTIVE_INNER_RADIUS,
-                outer_radius = self.ACTIVE_OUTER_RADIUS,
-                width = self.ACTIVE_LINE_WIDTH,
-                axis = self._rotation_axis,
-                color = ToolHandle.AllAxisSelectionColor
-            )
+        p_base1 = Vector(p_base0.x, p_base0.y + self.ARROW_HEAD_WIDTH, p_base0.z)
+        p_base2 = Vector(p_base0.x, p_base0.y - self.ARROW_HEAD_WIDTH, p_base0.z)
+        p_base3 = Vector(p_base0.x + self.ARROW_HEAD_WIDTH, p_base0.y, p_base0.z)
+        p_base4 = Vector(p_base0.x - self.ARROW_HEAD_WIDTH, p_base0.y, p_base0.z)
+        p_base5 = Vector(p_base0.x, p_base0.y, p_base0.z + self.ARROW_HEAD_WIDTH)
+        p_base6 = Vector(p_base0.x, p_base0.y, p_base0.z - self.ARROW_HEAD_WIDTH)
 
-            self.setSelectionMesh(mb.build())
+        mb.addFace(p_base1, p_head, p_base3, color=self.color)
+        mb.addFace(p_base3, p_head, p_base2, color=self.color)
+        mb.addFace(p_base2, p_head, p_base4, color=self.color)
+        mb.addFace(p_base4, p_head, p_base1, color=self.color)
+        mb.addFace(p_base5, p_head, p_base1, color=self.color)
+        mb.addFace(p_base6, p_head, p_base1, color=self.color)
+        mb.addFace(p_base6, p_head, p_base2, color=self.color)
+        mb.addFace(p_base2, p_head, p_base5, color=self.color)
+        mb.addFace(p_base3, p_head, p_base5, color=self.color)
+        mb.addFace(p_base5, p_head, p_base4, color=self.color)
+        mb.addFace(p_base4, p_head, p_base6, color=self.color)
+        mb.addFace(p_base6, p_head, p_base3, color=self.color)
 
-        self._arrow.buildMesh()
+        p_tail1 = Vector(p_tail0.x, p_tail0.y + self.ARROW_TAIL_WIDTH, p_tail0.z)
+        p_tail2 = Vector(p_tail0.x, p_tail0.y - self.ARROW_TAIL_WIDTH, p_tail0.z)
+        p_tail3 = Vector(p_tail0.x + self.ARROW_TAIL_WIDTH, p_tail0.y, p_tail0.z)
+        p_tail4 = Vector(p_tail0.x - self.ARROW_TAIL_WIDTH, p_tail0.y, p_tail0.z)
+        p_tail5 = Vector(p_tail0.x, p_tail0.y, p_tail0.z + self.ARROW_TAIL_WIDTH)
+        p_tail6 = Vector(p_tail0.x, p_tail0.y, p_tail0.z - self.ARROW_TAIL_WIDTH)
+
+        p_tail_base1 = Vector(p_base0.x, p_base0.y + self.ARROW_TAIL_WIDTH, p_base0.z)
+        p_tail_base2 = Vector(p_base0.x, p_base0.y - self.ARROW_TAIL_WIDTH, p_base0.z)
+        p_tail_base3 = Vector(p_base0.x + self.ARROW_TAIL_WIDTH, p_base0.y, p_base0.z)
+        p_tail_base4 = Vector(p_base0.x - self.ARROW_TAIL_WIDTH, p_base0.y, p_base0.z)
+        p_tail_base5 = Vector(p_base0.x, p_base0.y, p_base0.z + self.ARROW_TAIL_WIDTH)
+        p_tail_base6 = Vector(p_base0.x, p_base0.y, p_base0.z - self.ARROW_TAIL_WIDTH)
+
+        mb.addFace(p_tail1, p_tail_base1, p_tail3, color=self.color)
+        mb.addFace(p_tail3, p_tail_base3, p_tail2, color=self.color)
+        mb.addFace(p_tail2, p_tail_base2, p_tail4, color=self.color)
+        mb.addFace(p_tail4, p_tail_base4, p_tail1, color=self.color)
+        mb.addFace(p_tail5, p_tail_base5, p_tail1, color=self.color)
+        mb.addFace(p_tail6, p_tail_base6, p_tail1, color=self.color)
+        mb.addFace(p_tail6, p_tail_base6, p_tail2, color=self.color)
+        mb.addFace(p_tail2, p_tail_base2, p_tail5, color=self.color)
+        mb.addFace(p_tail3, p_tail_base3, p_tail5, color=self.color)
+        mb.addFace(p_tail5, p_tail_base5, p_tail4, color=self.color)
+        mb.addFace(p_tail4, p_tail_base4, p_tail6, color=self.color)
+        mb.addFace(p_tail6, p_tail_base6, p_tail3, color=self.color)
+
+        mb.addFace(p_tail_base1, p_tail_base3, p_tail3, color=self.color)
+        mb.addFace(p_tail_base3, p_tail_base2, p_tail2, color=self.color)
+        mb.addFace(p_tail_base2, p_tail_base4, p_tail4, color=self.color)
+        mb.addFace(p_tail_base4, p_tail_base1, p_tail1, color=self.color)
+        mb.addFace(p_tail_base5, p_tail_base1, p_tail1, color=self.color)
+        mb.addFace(p_tail_base6, p_tail_base1, p_tail1, color=self.color)
+        mb.addFace(p_tail_base6, p_tail_base2, p_tail2, color=self.color)
+        mb.addFace(p_tail_base2, p_tail_base5, p_tail5, color=self.color)
+        mb.addFace(p_tail_base3, p_tail_base5, p_tail5, color=self.color)
+        mb.addFace(p_tail_base5, p_tail_base4, p_tail4, color=self.color)
+        mb.addFace(p_tail_base4, p_tail_base6, p_tail6, color=self.color)
+        mb.addFace(p_tail_base6, p_tail_base3, p_tail3, color=self.color)
+
+        self.setSolidMesh(mb.build())
+
+        self._handle.buildMesh()
 
     def _onSelectionCenterChanged(self) -> None:
         pass
 
     def setCenterAndRotationAxis(self, center: Vector, rotation_axis: Vector, arrow_direction: Vector = None):
-        """Sets the center and the plane of the rotation handle, then rotates the load arrow to the desired direction"""
-        axis = self._rotation_axis.cross(rotation_axis)
-        angle = angleBetweenVectors(rotation_axis, self._rotation_axis)
+        axis = self._handle.rotation_axis.cross(rotation_axis)
+        angle = angleBetweenVectors(rotation_axis, self._handle.rotation_axis)
 
         if axis.length() < 1.e-3:
             axis = rotation_axis
@@ -82,16 +137,16 @@ class LoadHandle(ToolHandle):
         matrix = Quaternion.fromAngleAxis(angle, axis)
         self.rotate(matrix, SceneNode.TransformSpace.World)
 
-        self._rotation_axis = rotation_axis
+        self._handle.rotation_axis = rotation_axis
 
-        translation = center - self._center
-        self._center = center
+        translation = center - self._handle.center
+        self._handle.center = center
         self.setPosition(center)
 
-        self._arrow.start += translation
-        self._arrow.direction = matrix.rotate(self._arrow.direction)
+        self.start += translation
+        self.direction = matrix.rotate(self.direction)
 
-        self._arrow.rotateByVector(arrow_direction)
+        self.rotateByVector(arrow_direction)
 
     def setEnabled(self, enabled: bool):
         super().setEnabled(enabled)
@@ -102,3 +157,33 @@ class LoadHandle(ToolHandle):
         super().setVisible(visible)
         for child in self._children:
             child.setVisible(visible)
+
+    def setRotatorVisible(self, enabled: bool):
+        self._handle.setVisible(enabled)
+
+    def flip(self, normalAxis: Vector):
+
+        self.rotateByAngle(math.pi)
+
+        self.start = self.start + self.ARROW_TOTAL_LENGTH * self.direction
+        self.direction = -1 * self.direction
+
+        self.setPosition(self.start)
+
+    def rotateByVector(self, direction: Vector):
+        if direction is None:
+            return
+
+        if self.direction is None:
+            self.direction = direction
+            return
+
+        self.rotatebyAngle(
+            angleBetweenVectors(direction, self.direction)
+        )
+        self.direction = direction
+
+    def rotateByAngle(self, angle: float):
+        self.rotate(
+            Quaternion.fromAngleAxis(angle, self._rotation_axis)
+        )
